style: split socket IO helpers out of AutomationServer
The merge from main pushed AutomationServer.swift past SwiftLint's
file_length (403 > 400) and type_body_length (268 > 250) limits; the
stateless descriptor helpers now live in AutomationServer+SocketIO.swift.
